refactor(semaphore): update semaphore implementation for improved efficiency and new structure
This commit refactors the semaphore implementation by introducing a new structure and several optimizations. The maximum value for semaphores is now defined as 32767 instead of the previous maximum representation, which enhances clarity and consistency across implementations.

Key changes include:
- Replacing `int` with `short` in mutex and semaphore structures to save memory.
- Consolidating semaphore function implementations under a more streamlined design.
- Removing deprecated semaphore functions and associated code from the POSIX layer.
- Enhancing thread-safety and performance for both multithreaded and single-threaded environments.

These improvements aim to reduce memory usage and increase the efficiency of semaphore operations while maintaining backward compatibility where possible.

+++ include/pthread.h
@@ -46,19 +46,19 @@ typedef struct {
 typedef struct {
     void *owner;
     void *wait_queue;    
-    int locked;
-    int protocol;              /* Priority protocol */
-    int type;                  /* Mutex type */
-    int prioceiling;           /* Priority ceiling */
-    int saved_priority;        /* Saved priority for ceiling protocol */
-    int lock_count;            /* Lock count for recursive mutexes */
+    short locked;
+    short protocol;              /* Priority protocol */
+    short type;                  /* Mutex type */
+    short prioceiling;           /* Priority ceiling */
+    short saved_priority;        /* Saved priority for ceiling protocol */
+    short lock_count;            /* Lock count for recursive mutexes */
 } pthread_mutex_t;
 
 typedef struct {
-    int type;                 /* Mutex type: NORMAL, RECURSIVE, ERRORCHECK */
-    int pshared;              /* Process-shared flag */
-    int protocol;             /* Priority protocol */
-    int prioceiling;          /* Priority ceiling value */
+    short type;                 /* Mutex type: NORMAL, RECURSIVE, ERRORCHECK */
+    short pshared;              /* Process-shared flag */
+    short protocol;             /* Priority protocol */
+    short prioceiling;          /* Priority ceiling value */
 } pthread_mutexattr_t;
 
 /* Condition variable */
